# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username or name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        avatarColor:
          type: string
        status:
          type: string
          enum: [online, offline, away]
        statusText:
          type: string
          nullable: true
        lastSeen:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        isBot:
          type: boolean
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        showOnlineStatus:
          type: boolean
        readReceiptsEnabled:
          type: boolean
        totpEnabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - avatarColor
        - status
        - isBot
        - isVerified
        - isAdmin
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        avatarColor:
          type: string
        statusText:
          type: string
        status:
          type: string
          enum: [online, offline, away]
        showOnlineStatus:
          type: boolean
        readReceiptsEnabled:
          type: boolean
        statusAnimation:
          type: string
...